v2 added Favorite and Purchase capabilities

+++ src/main/java/com/bookstore/bookstore/book/BookRepository.java
@@ -1,16 +1,9 @@
 package com.bookstore.bookstore.book;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
 import java.util.List;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
     List<Book> findByTitleContainingIgnoreCaseOrAuthorContainingIgnoreCase(String keyword, String keyword1);
-
-//    @Query("SELECT b FROM Book b WHERE LOWER(b.title) LIKE LOWER(CONCAT('%', :keyword, '%')) " +
-//            "OR LOWER(b.author) LIKE LOWER(CONCAT('%', :keyword, '%'))")
-//    List<Book> searchByTitleOrAuthor(@Param("keyword") String keyword);
 
 }
